--------Additions-------- MPlayerController Class: Holds all of the needed delegates that allow the player to interact with objects in the world and switch pages in the journal
LoreManager Class: Initializes all of the lore tablets with lore that comes from a data table.

EventZone Class: These are Trigger Boxes that can be placed around the map to determine what level of event could happen in a specific zone.

Glowstick Class:  Glowsticks are used by the player to light his/her surroundings and give them a better chance of not getting lost.

CrawableArea Class: Will be placed where the player can crawl, this class will limit the camera movement of the player to add a more immersive experience.

LoreTablet Class: Lore tablets are placed around the map and can be interacted with. When interacted with the player will gain the knowledge on the tablets and write it in their journal for future viewing.

ScarePoint Class: This actor is placed around the world and acts as a positional scare point that also has its own scare overrides.

WInteraction Widget Class: Handles all the functionality that allows players to interact with scare tablets and other objects if needed.

WJournal Widget Class: Handles all the functionality for the player to open his/her journal and go through the pages.

                            -------- Changes--------

Map Changes: The Demo map is currently being blocked out and demo props are being added.

Quixel Bridge: This plugin has been imported and is planned to be used for future-level design and the demo.

PlayerCharacter Class: This class has been changed to account for multiplayer. This includes flashlight/glowstick replicating properly. In addition, crawling functionality had been added too.

CustomAnimInstance Class: Now handles crouch/crawl animations and transitions.

MisfortunateGameInstance Class: Moved the scare event manager singleton onto the GameMode to allow better multiplayer capabilities.

MisfortunateGameMode Class: Increased functionality to the scare system by adding a function to select characters to scare on the server instead of on the client. In addition to multiple miner getter/setter functions

ScareEventManager Class: The TMap for scares has recently been recreated with more functionality and allows for settings to be changed for each scare. In Addition, this class will now account for and use the ScarePoints are placed within the map.